refactor: refatorando seção de categorias
